Implement echo client
We also made some small changes to the server code--notably, changing
how we test when the input from the socket is finished. (The old version
did the "is the input finished" check and "read from the input" step
separately, and we were worried there might be race conditions if, eg,
the input was closed between our check and our read.)

The lab should be all finished and working at this point.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>